Setuped the architecture a bit. Made some small changes to the product component files, added parent and child routes. Updated the Breadcrumbs component a bit, there is a connection between the 'get-product' component and the breadcrumb service, so ill have to think of a way to separate these two, but thats gonna be it for now. The main-content component will have to be separated from the side navigation, ill look for that in the near feature.

+++ src/Bigisoft.Ecommerce.Server/Features/Products/Commands/DeleteProductCommand.cs
@@ -1,4 +1,3 @@
-using Azure.Core;
 using Bigisoft.Ecommerce.Server.Infrastructure.Data;
 using MediatR;
 
@@ -8,13 +7,17 @@ public sealed record DeleteProductCommand(int Id) : IRequest<bool>;
 
 public sealed class DeleteProductCommandHandler(EcommerceDbContext context) : IRequestHandler<DeleteProductCommand, bool>
 {
-    public async Task<bool> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
+    public async Task<bool> Handle(DeleteProductCommand? request, CancellationToken cancellationToken)
     {
-        var product = await context.Products.FindAsync(request.Id);
+        if (request != null)
+        {
+            var product = await context.Products.FindAsync(request.Id, cancellationToken);
         
-        if (product == null) return false;
+            if (product == null) return false;
 
-        context.Products.Remove(product);
+            context.Products.Remove(product);
+        }
+
         return await context.SaveChangesAsync(cancellationToken) > 0;
     }
 }
